Moved the service factory to the publisher handler namespace
